refactor: move helpers/command to helpers/parse, split ast.py by domain
The package was originally created for cmd task parsing but now covers
three distinct domains. Rename and split accordingly:

- helpers/parse/core.py    — parser infrastructure (was ast_core.py)
- helpers/parse/command.py — bash/cmd AST nodes (was ast.py)
- helpers/parse/template.py — template string nodes (extracted)
- helpers/parse/envfile.py  — envfile AST nodes (extracted)

Public API (parse_poe_cmd, resolve_template, resolve_command_tokens)
and all call sites updated. No logic changes.
